# Make `buffer.toString("base64")` 4x faster