--- a/bench/snippets/base64-buffer-to-string.mjs
+++ b/bench/snippets/base64-buffer-to-string.mjs
@@ -1,0 +1,14 @@
+import { bench, run } from "./runner.mjs";
+import { Buffer } from "node:buffer";
+
+const bigBuffer = Buffer.from("hello world".repeat(10000));
+const converted = bigBuffer.toString("base64");
+bench("Buffer.toString('base64')", () => {
+  return bigBuffer.toString("base64");
+});
+
+// bench("Buffer.from(str, 'base64')", () => {
+//   return Buffer.from(converted, "base64");
+// });
+
+await run();