This adds https://github.com/aklomp/base64 as a dependency to Bun, which is the same library that Node uses.

```zig
cpu: Apple M1 Max
runtime: bun 0.6.13_debug (arm64-darwin)

benchmark                      time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------- -----------------------------
Buffer.toString('base64')   40.41 µs/iter    (12.5 µs … 10.22 ms)  15.58 µs 561.83 µs   1.13 ms

❯ bun
cpu: Apple M1 Max
runtime: bun 0.6.12 (arm64-darwin)

benchmark                      time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------- -----------------------------
Buffer.toString('base64')  269.96 µs/iter     (177 µs … 38.42 ms) 186.92 µs 905.17 µs   3.41 ms

❯ node
cpu: Apple M1 Max
runtime: node v20.1.0 (arm64-darwin)

benchmark                      time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------- -----------------------------
Buffer.toString('base64')  129.44 µs/iter    (22.08 µs … 6.61 ms)  57.79 µs   1.65 ms   2.02 ms


❯ deno
cpu: unknown
runtime: deno 1.34.2 (aarch64-apple-darwin)

benchmark                      time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------- -----------------------------
Buffer.toString('base64')   154.7 µs/iter   (36.29 µs … 12.95 ms) 105.54 µs   1.44 ms   1.91 ms
```